# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts/{postId}/tags/{tagId}": {
      "post": {
        "operationId": "PostController_AddTag",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "409": {
            "description": "Conflict."
          }
        },
        "summary": "Create tag.",
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_RemoveTag",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfull."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Delete tag.",
        "tags": [
          "Post"
        ]
      }
    },
    "/tags": {
      "get": {
        "operationId": "TagsController_getTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfull."
          }
        },
        "summary": "Get tags.",
        "tags": [
          "Tags"
        ]
      },
      "post": {
        "operationId": "TagsController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Create tag.",
        "tags": [
          "Tags"
        ]
      }
    },
    "/tags/{id}": {
      "patch": {
        "operationId": "TagsController_updateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Tag not found."
          },
          "409": {
            "description": "Conflict."
          }
        },
        "summary": "Update tag.",
        "tags": [
          "Tags"
        ]
      },
      "delete": {
        "operationId": "TagsController_deleteTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfull."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Tag not found."
          }
        },
        "summary": "Delete tag.",
        "tags": [
          "Tags"
        ]
      }
    },
...
    "/users": {
      "get": {
        "operationId": "UserController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}/follow": {
      "post": {
        "operationId": "UserController_followUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Conflict."
          }
        },
        "summary": "Follow User.",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_unfollowUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "User not found or not currently following."
          }
        },
        "summary": "Unfollow User.",
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}/followers": {
      "get": {
        "operationId": "UserController_getUserFollowers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          }
        },
        "summary": "Get Follower User.",
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}/following": {
      "get": {
        "operationId": "UserController_getUserFollowings",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          }
        },
        "summary": "Get Following User.",
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blog API",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTagDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}